Rename the ocean module to explainer Rename ocean module to explainer with MIPExplainer as part of mip file.
