+++ ocean/explainer/__init__.py
@@ -1,0 +1,3 @@
+from .mip import MIPExplainer
+
+__all__ = ["MIPExplainer"]
