--- ocean/ocean/__init__.py
@@ -1,3 +1,0 @@
-from .ocean import MIPExplainer
-
-__all__ = ["MIPExplainer"]
